requirements.txt
pandas==1.4.2
python-docx==0.8.11
openpyxl==3.0.9
python-dotenv==0.20.0
requests==2.27.1
Flask==2.1.2
Pillow==9.1.1